+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Enable blob data soft delete on Azure.
+
 ## [0.10.0] - 2025-01-07
 
 ### Changed